Set up classroom selector to use vuex
