+++ kolibri/plugins/management/assets/src/vuex/actions.js
@@ -1,0 +1,7 @@
+function addClassroom({ dispatch }, attrs) {
+  dispatch('ADD_CLASSROOM', attrs);
+}
+
+export {
+  addClassroom,
+};
