+++ kolibri/plugins/management/assets/src/vuex/getters.js
@@ -1,0 +1,3 @@
+export function getClassrooms(state) {
+  return state.classrooms;
+}
