+++ package.json
@@ -74,6 +74,7 @@
     "vue-html-loader": "^1.0.0",
     "vue-loader": "indirectlylit/vue-loader#pad-css",
     "vue-style-loader": "^1.0.0",
+    "vuex": "0.6.3",
     "webpack": "^1.12.12",
     "webpack-bundle-tracker": "0.0.93",
     "webpack-dev-middleware": "^1.4.0",
